refactor(calc): パッケージを整えた
