+++ app/src/main/java/com/example/yosuke/calculator/di/modules/MainActivityModule.kt
@@ -1,4 +1,4 @@
-package com.example.yosuke.calculator.di
+package com.example.yosuke.calculator.di.modules
 
 import com.example.yosuke.calculator.MainActivity
 import dagger.Module
